requirements.txt
# Core
mistralai>=1.0.0
mistral-common[sentencepiece]>=1.0.0
python-dotenv>=1.0.0
tenacity>=8.0.0

# PDF Processing
pymupdf>=1.23.0

# ML
numpy>=1.24.0
torch>=2.0.0
openreview-downloader
openreview-py>=1.0.0

# CRAG - LangGraph and LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-community>=0.3.0

# Pinecone for persistent hybrid search
pinecone>=5.0.0
pinecone-text>=0.9.0

# Web Search
tavily-python>=0.5.0

# Pydantic for structured outputs
pydantic>=2.0.0